# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  controller.TypeNeighbor:
    properties:
      hostname:
        type: string
      ip:
        type: string
    type: object
  controller.TypeNeighborInfo:
    properties:
      code:
        type: integer
      info:
        additionalProperties: true
        type: object
    type: object
  controller.TypeNeighborInfos:
    properties:
      neighbors:
        additionalProperties:
          $ref: '#/definitions/controller.TypeNeighborInfo'
        type: object
    type: object
  controller.TypeNeighbors:
    properties:
      neighbors:
        items:
          $ref: '#/definitions/controller.TypeNeighbor'
        type: array
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/neighbor:
    delete:
      consumes:
        - application/x-www-form-urlencoded
      description: DeleteNeighbor.
      parameters:
        - description: Neighbor IP
          in: formData
          name: ip
          required: true
          type: string
        - description: Neighbor Hostname
          in: formData
          name: hostname
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TypeNeighbor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: DeleteNeighbor
      tags:
        - API
        - CUBE
    get:
      consumes:
        - application/x-www-form-urlencoded
      description: GetNeighbor.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TypeNeighbors'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: GetNeighbor
      tags:
        - API
        - CUBE
    post:
      consumes:
        - application/x-www-form-urlencoded
      description: PutNeighbor.
      parameters:
        - description: Neighbor IP
          in: formData
          name: ip
          required: true
          type: string
        - description: Neighbor Hostname
          in: formData
          name: hostname
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TypeNeighbors'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: PutNeighbor
      tags:
        - API
        - CUBE
    put:
      consumes:
        - application/x-www-form-urlencoded
      description: PutNeighbor.
      parameters:
        - description: Neighbor IP
          in: formData
          name: ip
          required: true
          type: string
        - description: Neighbor Hostname
          in: formData
          name: hostname
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TypeNeighbors'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: PutNeighbor
      tags:
        - API
        - CUBE
  /cube/neighbor/info:
    get:
      consumes:
        - application/x-www-form-urlencoded
      description: GetNeighbor.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TypeNeighborInfos'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: GetNeighbor
      tags:
        - API
        - CUBE
...
schemes:
  - http
  - https
securityDefinitions:
  None:
    type: basic
swagger: "2.0"